+++ dart/gui/osg/ShapeFrameNode.cpp
@@ -48,6 +48,7 @@
 #include "dart/gui/osg/render/MeshShapeNode.hpp"
 #include "dart/gui/osg/render/SoftMeshShapeNode.hpp"
 #include "dart/gui/osg/render/LineSegmentShapeNode.hpp"
+#include "dart/gui/osg/render/PointCloudShapeNode.hpp"
 #if HAVE_OCTOMAP
 #include "dart/gui/osg/render/VoxelGridShapeNode.hpp"
 #endif
@@ -67,6 +68,7 @@
 #include "dart/dynamics/MeshShape.hpp"
 #include "dart/dynamics/SoftMeshShape.hpp"
 #include "dart/dynamics/LineSegmentShape.hpp"
+#include "dart/dynamics/PointCloudShape.hpp"
 #if HAVE_OCTOMAP
 #include "dart/dynamics/VoxelGridShape.hpp"
 #endif
@@ -288,6 +290,15 @@ void ShapeFrameNode::createShapeNode(
         std::dynamic_pointer_cast<LineSegmentShape>(shape);
     if(lss)
       mRenderShapeNode = new render::LineSegmentShapeNode(lss, this);
+    else
+      warnAboutUnsuccessfulCast(shapeType, mShapeFrame->getName());
+  }
+  else if(PointCloudShape::getStaticType() == shapeType)
+  {
+    std::shared_ptr<PointCloudShape> lss =
+        std::dynamic_pointer_cast<PointCloudShape>(shape);
+    if(lss)
+      mRenderShapeNode = new render::PointCloudShapeNode(lss, this);
     else
       warnAboutUnsuccessfulCast(shapeType, mShapeFrame->getName());
   }